Add ActivityLiveCycle to monitor MainActivity display state, to decide send/keep  the GeoFencing intent to MainActivity

+++ app/src/main/java/btremote999/geofencetest/utils/Common.java
@@ -19,7 +19,9 @@ public class Common {
         }
         else {
             // patch case for samsung s7 with xxxhdp
-            return (int) (dp * 4);
+            return dp * 4;
         }
     }
+
+
 }
